Improved point blank hit check

+++ razor.h
@@ -40,8 +40,8 @@ public:
     void Hit(float damage, int ownerID);
     void Set(Vector3 position);
 protected:
-    double topSpeed_;
-    double aimSpeed_;
+    float topSpeed_;
+    float aimSpeed_;
 
     Node* topNode_;
     Node* bottomNode_;
